requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
decorator==4.4.2
importlib-metadata==3.7.2
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
PyYAML==5.4.1
retry==0.9.2
selenium==3.141.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
zipp==3.4.1
